frontend: rename and cleanup socket types
This is a cleanup:
- merges socket related types into transport-common.ts
- moves the transport code of websocket.ts to transport-websocket
- keeps the logic of which transport to use in websocket.ts
- consistently uses T prefix
